bloom: add a 'complete' method.
not sure I like the name yet.  basically the idea is that we'd
like to be able to test for the unlikely-but-not-impossible case
that the bitvector completely fills up and thus we have exhausted
all possible configurations.

this is silly now with 65535 bits, but in the future we probably
just want to have 'n' classes of bits, and for hash fqn collision
reasons multiply that 'n' by some constant for the total number of
bits.  in that case, we might end up with significantly fewer
bits, and the /whole point/ is that we'd like to exhaustively test
all paths in finite time.

the end goal being that every Value has its own bloom filter
associated with it, and we try the program where each Value takes
a value from every possible class.