Start of a revised 'Variable' hierarchy.
That actually splits things into proper classes instead of just
using a bloom filter.  We may want to replace the ind_sample()
calls inside the TC::value() methods with bloom filters in the
future, with the idea of sampling intelligently within each class,
but probably best to ignore that entirely for a first pass.

Rust specialization might make this nicer in the future.