Mostly a bunch of comments documenting what each thing is.
Because I confused even myself.  The problem is that 'Free' was
invented to do a lot more and probably take a role in code
generation in the future, but right now it basically is just doing
what 'Value' does due to incomplete implementations.  Make it
clear what each is for.

But I also added a TypeClass for Enums, interesting because it is
actually dependent on the input API due to its need to get the
values from the enum in the defined API.  I suspect user-defined
types will go a similar route.