Large rework of variables.
This implements a 'gen' function that takes an "API", which is
just a collection of functions along with variable (dataflow)
information.  Each variable has a 'Value' associated with it now,
so that we can generate values of the appropriate class on demand.

Move Type's "Name" implementation to the file with Types, so that
it has those everywhere.

Also, get rid of all the lifetime specifiers and just use value
types.  It's too painful to get working and not worth it.