Iterate through depvars to print out arguments.
This is hackish and gross.  A better approach would be to attach
the variable information, free or dependent, to the list of
arguments in the function.  Then we don't have this convoluted
method of figuring out which variables are associated with which
functions.

Even worse, this doesn't capture free variables at the moment
because the free variables don't have enough information to
associate them to a function parameter.