Total rewrite of the variable and Function logic.
The old way was too difficult to (dis)associate variables with
functions.  There are times when we want them separate (i.e.
iterate through every 'free' variable we'll have so that we can
declare it) and then times when we need to have just that
association but metadata along with that association.

This basically gets rid of variables as separate entities.
Functions are things and they have a number of arguments; they
also have a number of slots for "variable Source"s.  That slot
holds not just the variable information but the metadata (e.g.
deref or not) needed to apply that variable at that particular
location.

Removed all the code that used the old interface, so with this
commit we are not compiling+executing anymore ...