Remove a bunch of code that we won't end up using.
They could have been extended out to do the more traditional
fuzzing situation of having extremely large sets of values we want
to represent and we want to sample through that space
intelligently.  But we've moved more to an idea of having a rather
small finite set of states, so they won't be useful for us even in
the future.