Start implementing a new, struct-based Source.
The enum-based one cannot work, because enums are immutable by
definition, so we don't be able to iterate through our Generators.

Further, it seems we need RefCell for the interior mutability.  We
seem to need vectors because there is no way to denote an "empty"
Rc or RefCell, too... hoping I find a workaround for that sometime
in the future.