Fix iteration to also include/consider return values.
Not 100% sure the utility of this.  We are currently initializing
return values with the function itself, so any potential Generator
associated with it never queried anyway.  Are there cases where we
would want to generate return values?  Should we simply enforce
that return values cannot be given 'free' variables?  We do have a
'retval' class for them ...