Rework CString generator, and TypeClasses in general.
Made the TypeClass::value mutable, so that it could store a random
number generator instead of creating one every time.  Not very
happy with that: seems like Rust really needs C++'s "mutable" for
exactly a case like this, the "I know I said I wouldn't change
this, but I promise, this change is benign, really" case.

Anyway prove the concept works by adding type classes for
printable and non-printable characters, and switch up the CString
generator to use those type classes.

I also made the CString generator return NULL in its 0 case, as
it is designed to do, though this in turn breaks the 'hash' test
case that is builtin.  That program needs real strings, so both
the 0 and 1 case need to somehow be disabled ... but we don't have
any mechanism for disabling generator states at present.