Implement user-defined generators.
This adds parsing and interpretation for generators that are
stored in configuration files instead of hand-coded into the Rust
code itself.  It does not yet change any of the default test case
to utilize that machinery; more just a proof of concept in this
commit.

This is needed because we cannot possibly predict the user's
Generator needs.  For example, an API that takes a CString might
actually only accept certain characters---think fopen's mode bits
for example.