Rework generator grammar to allow "std::???" names.
Basically by special casing the "std:" prefix in the grammar,
pretty ugly.  It seems that adding in the ":" character to any
part of LName would then mess up *other* states later, by suddenly
making the ":" be a word separator instead of accepting e.g. the
"state:digit" syntax.

Fortunately we weren't actually using some of the digits we had in
there anyway, so it was simplest to just remove them.