Merge ctime and loaded generators, and lookup from there.
We now have some generators runtime-loaded from stdgen.hf and some
generators that are hardcoded in at compile-time.  Merge the two
into one big Vec<Generator>, and then use that to instantiate
generators ad hoc based on the name instead of hardcoding in the
generator type.

This necessitated giving all generators, even compile-time ones,
names.  It also means we needed the ability to copy a generator,
which was generally difficult because a) it's a trait, and b) we
always work with them in boxes (... because it's a trait).

In main we still hardcode too much of the 'hash' case as types and
generator types.  We need to convert UDTs and Enums to be
something that have unique names.