Merge 'api' and main type system.
This reworks typ::Type a bit so that it mirrors how we parse
things a bit better.

We still have a mirror type system for parsing that uses more
strings for references to some nebulous type.  This allows us to
quickly build a parse tree without having to do all sorts of type
lookups and type definitions from within the parse, meaning the
parsing code can stay pretty simple.  As it is autogenerated, it's
desirable that we keep it as simple as possible.