Add a "clone" for sources.
In particular this lets us get at a non-Option'd Rc'd RefCell'd
copy, which is so much easier to work with.
Should probably make a new type instead of directly returning an
Rc<RefCell<Source>> from the creation methods, and then have the
new type auto-borrow / auto-deref to what we need, because this
stuff is a nightmare.