func_from_decl half-implementation and test.
It will need to be revisited a bit.  The problems are that we have
lost the name by the time we get into a FuncDecl, and that we
don't have any useful generators at the time.  Seems like we're
going to need to also pass the current list of generators to these
*_from_decl functions...