Simpler typing for UDT parsing, and type resolution.
Doing proper type resolution for a struct.  Still missing out on
generators though; when the user defines a 'struct x{}', we should
add a "default" generator for 'x' at the very least.

We might have an issue with generators for user-defined types:
type resolution needs the list of defined generators.  this means
generators need to be defined before types, but a user-defined
obviously cannot be defined into we get to parsing the user
specification.  hrm.