make Generator::value() immutable.
... and some corresponding functions.  If we just reference
rand::thread_rng() dynamically as needed instead of storing it in
the object, then nothing about the object changes when we need a
random number.  then all these things can be immutable, which they
really are: even though the exact value can (in some cases)
change, the *class* of the values that a generator gives does not
change between multiple invocations of value().