Start cleaning up Function::Arguments.
To not have the variable::Source directly, but instead just take
an expression and use that for everything.  An expression, of
course, can reference a variable, so it's essentially the same.

I replicated a bunch of the Generator control stuff to Argument,
which cleaned a lot of things up and hid all the scary deref_mut
and runtime borrows and the like.  Using expressions also
simplified the main code a bit, notably by dropping rvalue() since
it is basically Expression::codegen().