Temporarily introduce a "SimpleSym" expression.
We want expressions to reference api::Symbols instead of
variable::Sources.  However, an outright switchover is too much
effort.  This is an intermediate step that introduces the end
result we want in parallel with the deprecated way.

Further commits will switch more code to use the SimpleSym
interface, and eventually we will drop what is "Simple" now and
rename "SimpleSym" to just "Simple".