Add String-based statements and expressions.
And have the parsing phase generate those instead of "real"
expressions.  We can't really create real expressions at that
point, because a real expression references the symbol table and
we have not yet built the symbol table while we're parsing.  So
generate basically a bunch of references and then during analysis
move all those references to real actual Expressions/Statements.