Get rid of 'Source'!  Finally.
Drops 'Source'.  This was a holdover from the early days, where I
though we wouldn't need a "real" language under the hood.  We do
need a real language.  And since we have one, we represent all the
information and more via variable declarations and Expressions.

This is also nice because Rc<RefCell> nonsense has always been a
huge pain, and the 'real language' API is much cleaner in that it
doesn't need that crap.