add new 'constraint' statement.
Constraints take a boolean expression.  If that expression
evaluates to false, the test is not run.  This allows one to avoid
testing programs for which a combination of generators needs to be
true to validate the test.  For example, one might have a
constraint of 'dims[0] > 0 && dims[1] > 0 && dims[2] > 0' to
validate the dimensions of a regular grid are positive/make sense.