add constants for parsinadd constants to parse tree.
we should really rework how this is done.  these are basically
just copies of the stmt::* variants; instead we should pass the
strings directly and then we can do the conversion when we build
the Statements and Expressions.  Long-term, that'd be better for
error reporting in particular, and in the short term it more
importantly keeps the code within the grammar very simple,
complexity is done later.