Start of a "single" generator, generically.
*All* variables in our programs have generators attached.  But
sometimes that doesn't make sense.  Let's say we are fuzzing the
program:
   foo = pow(bar, baz);
   verify:new(foo >= 0);
In this case, both "bar" and "baz" are free variables that should
take on a number of different values.  But "foo"'s generated value
is not relevant, because it will be immediately overwritten as the
result of the function call.

Thus "foo" should get one of these new SingleGen<T>erators, and
thereby we can still "generate" a value for it using the same
normal variable machinery, and because it only has a single state
it will not needlessly explode the fuzzspace.