SingleGen: get rid of the argument to create.
i.e. the client does not get to set the value, it just uses the
default for the type that is being generated.

Unfortunately we cannot get rid of the field entirely, because
Rust insists that a struct's template parameter must be used for a
field within the struct, whereas we only ever use it inside method
definitions.