Add a new 'decl' to Generators.
This will allow us to switch things up so that declaration and
setting to the initial value() can be distinct pieces of code.
We'll need this because we need to expand out the idea of value()
to be able to include multiple lines.  A 'value' should in the
future set some bit of state on the underlying data structure,
with any single instance of a program doing this multiple times.

Think of defining a graph with a set of children.  We want decl
for:
	node n;
and then value() might:
	n.set_bit_foo();
or:
	node chld;
	n.add_child(chld);
and we might have a program that combines these an arbitrary
number of "value()" settings:
	node n;
	node chld;
	n.add_child(chld);
	n.set_bit_foo();
	n.set_bit_bar();
	n.set_baz("foobar");

For that scheme, we can't have the declaration get mixed up with
some operation that mutates a node, because then we'll never be
able to generate any programs which do *not* perform that
operation.  i.e. if state 0 was "set_bit_foo()", we'd never be
able to create tests where foo was not set.